+++ spec/integration/braintree/customer_spec.rb
@@ -456,7 +456,7 @@ describe Braintree::Customer do
           }
         )
         result.success?.should == true
-        result.customer.credit_cards.first.venmo_sdk?.should == true
+        result.customer.credit_cards.first.venmo_sdk?.should == false
       end
     end
 
@@ -973,7 +973,12 @@ describe Braintree::Customer do
 
     it "returns associated us bank accounts" do
       result = Braintree::Customer.create(
-        :payment_method_nonce => generate_valid_us_bank_account_nonce
+        :payment_method_nonce => generate_non_plaid_us_bank_account_nonce,
+        :credit_card => {
+          :options => {
+            :verification_merchant_account_id => SpecHelper::UsBankMerchantAccountId,
+          }
+        }
       )
       result.should be_success
 
@@ -984,9 +989,9 @@ describe Braintree::Customer do
       us_bank_account = found_customer.us_bank_accounts.first
       us_bank_account.should be_a(Braintree::UsBankAccount)
       us_bank_account.routing_number.should == "021000021"
-      us_bank_account.last_4.should == "1234"
+      us_bank_account.last_4.should == "0000"
       us_bank_account.account_type.should == "checking"
-      us_bank_account.account_holder_name.should == "Dan Schulman"
+      us_bank_account.account_holder_name.should == "John Doe"
       us_bank_account.bank_name.should =~ /CHASE/
     end