+++ spec/spec_helper.rb
@@ -39,7 +39,8 @@ unless defined?(SPEC_HELPER_LOADED)
     ThreeDSecureMerchantAccountId = "three_d_secure_merchant_account"
     FakeAmexDirectMerchantAccountId = "fake_amex_direct_usd"
     FakeVenmoAccountMerchantAccountId = "fake_first_data_venmo_account"
-    UsBankMerhantAccountId = "us_bank_merchant_account"
+    UsBankMerchantAccountId = "us_bank_merchant_account"
+    AnotherUsBankMerchantAccountId = "another_us_bank_merchant_account"
     IdealMerchantAccountId = "ideal_merchant_account"
 
     TrialPlan = {